chore: restore hooks directory
